iOS connector server (#15)
Because of restrictions in iOS 17 on physical devices in terms of connections being made to hosts in the local network, we decided to turn the connection mechanism around. Instead of running a server socket on the automation host side and a client on the app side, we now run the server socket inside the connector and connect to from the automation host side.
This comes with pros and cons, but overall it seems to be an improvement.
The port can be configured via the `-port` launch argument.

* Try to change banner format

* Changed connector to be a server instead of a client

* New header format, increase protocolVersion to 180

* Fix datatype warning

* Clean up

* Fixed missing logs

* added comment

* changed delay to 1ms

* removed delay. Moved all code to main queue

* Show/Hide alert on disConnect/connect

* Better messaging

* Update readme.md